refacto: split boreal-py impl into multiple modules
